qr/requirements.txt
qrcode[pil]
Pillow
pyzbar